requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.4
django-admin-rangefilter==0.11.0
django-ckeditor==6.7.0
django-import-export==3.2.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
et-xmlfile==1.1.0
filelock==3.12.2
lxml==4.9.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
python-docx==0.8.11
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.3
xlrd==2.0.1
xlwt==1.3.0
